Added Popup Dialog on delete course, Search, Flex layout

+++ src/app/courses/pages/courses/courses.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Course } from '../../../shared/interfaces/course';
+import { Course } from '../../../shared/interfaces';
 import { CoursesService } from '../../../shared/services/courses.service';
 
-import { MatSnackBar } from '@angular/material';
+import { MatSnackBar, MatDialogRef, MatDialog } from '@angular/material';
+import { ConfirmDialogComponent } from '../../../shared/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -15,9 +16,13 @@ export class CoursesComponent implements OnInit {
   countAll: number;
   numStartItem = 0;
   countItems = 2;
+  searchStr = '';
+  dialogRef: MatDialogRef<ConfirmDialogComponent>;
 
-  constructor(private coursesService: CoursesService,
-    public snackBar: MatSnackBar
+  constructor(
+    private coursesService: CoursesService,
+    private dialog: MatDialog,
+    private snackBar: MatSnackBar,
   ) { }
 
   ngOnInit() {
@@ -26,7 +31,7 @@ export class CoursesComponent implements OnInit {
   }
 
   loadMore() {
-    this.coursesService.find(this.numStartItem, this.countItems).subscribe(
+    this.coursesService.find(this.searchStr, this.numStartItem, this.countItems).subscribe(
       resp => {
         this.countAll = resp.all;
         resp.items.map(course => this.courses.push(course));
@@ -35,12 +40,29 @@ export class CoursesComponent implements OnInit {
     this.numStartItem += this.countItems;
   }
 
+  searchCourses(searchStr: string) {
+    this.searchStr = searchStr;
+    this.numStartItem = 0;
+    this.courses = [];
+    this.loadMore();
+  }
+
   delete(course: Course) {
-    // TODO: Subscribe + get Count from Service answer
-    this.coursesService.delete(course);
-    this.countAll--;
-    this.courses = this.courses.filter(item => item !== course);
-    this.snackBar.open("This is the SnackBar Message", "Delete it!");
+    this.dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      disableClose: false
+    });
+    this.dialogRef.componentInstance.confirmMessage = "Are you sure you want to delete course?"
+
+    this.dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.coursesService.delete(course);
+        this.countAll--;
+        this.courses = this.courses.filter(item => item !== course);
+
+        this.snackBar.open("Course deleted.", "", {duration: 4000});
+      }
+      this.dialogRef = null;
+    });
   }
 
 }
